+++ lib/legion/extensions/chef/runners/cookbook.rb
@@ -5,6 +5,14 @@ module Legion
     module Chef
       module Runners
         module Cookbook
+          def list_cookbooks(**)
+            { cookbooks: connection(**).get('cookbooks').body }
+          end
+
+          def get_cookbook(name:, **)
+            { cookbook: connection(**).get("cookbooks/#{name}").body }
+          end
+
           include Legion::Extensions::Helpers::Lex
         end
       end