+++ lib/legion/extensions/chef/runners/data.rb
@@ -5,6 +5,14 @@ module Legion
     module Chef
       module Runners
         module Data
+          def list_data_bags(**)
+            { data_bags: connection(**).get('data').body }
+          end
+
+          def get_data_bag(name:, **)
+            { data_bag: connection(**).get("data/#{name}").body }
+          end
+
           include Legion::Extensions::Helpers::Lex
         end
       end