+++ lib/legion/extensions/chef/runners/group.rb
@@ -5,6 +5,14 @@ module Legion
     module Chef
       module Runners
         module Group
+          def list_groups(**)
+            { groups: connection(**).get('groups').body }
+          end
+
+          def get_group(name:, **)
+            { group: connection(**).get("groups/#{name}").body }
+          end
+
           include Legion::Extensions::Helpers::Lex
         end
       end