+++ lib/legion/extensions/chef/runners/license.rb
@@ -5,6 +5,10 @@ module Legion
     module Chef
       module Runners
         module License
+          def get_license(**)
+            { license: connection(**).get('license').body }
+          end
+
           include Legion::Extensions::Helpers::Lex
         end
       end