+++ lib/legion/extensions/chef/runners/policy.rb
@@ -5,6 +5,14 @@ module Legion
     module Chef
       module Runners
         module Policy
+          def list_policies(**)
+            { policies: connection(**).get('policies').body }
+          end
+
+          def get_policy(name:, **)
+            { policy: connection(**).get("policies/#{name}").body }
+          end
+
           include Legion::Extensions::Helpers::Lex
         end
       end