+++ lib/legion/extensions/chef/runners/search.rb
@@ -5,6 +5,11 @@ module Legion
     module Chef
       module Runners
         module Search
+          def search(index:, query: '*:*', rows: 1000, start: 0, **)
+            resp = connection(**).get("search/#{index}", { q: query, rows: rows, start: start })
+            { total: resp.body['total'], rows: resp.body['rows'] }
+          end
+
           include Legion::Extensions::Helpers::Lex
         end
       end