+++ lib/legion/extensions/chef/runners/user.rb
@@ -5,6 +5,14 @@ module Legion
     module Chef
       module Runners
         module User
+          def list_users(**)
+            { users: connection(**).get('users').body }
+          end
+
+          def get_user(name:, **)
+            { user: connection(**).get("users/#{name}").body }
+          end
+
           include Legion::Extensions::Helpers::Lex
         end
       end